+++ app/api/v1/get-token/route.ts
@@ -1,0 +1,26 @@
+import { validateRequest } from "@/lib/auth/helper";
+import streamServerClient from "@/lib/stream";
+
+export const dynamic = 'force-dynamic';
+export async function GET() {
+  try {
+    const { user } = await validateRequest();
+    if (!user) {
+      return Response.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
+    const expirationTime = Math.floor(Date.now() / 1000) + 60 * 60;
+
+    const issuedAt = Math.floor(Date.now() / 1000) - 60;
+
+    const token = streamServerClient.createToken(
+      user.id,
+      expirationTime,
+      issuedAt,
+    );
+    return Response.json({ token });
+  } catch (error) {
+    console.error(error);
+    return Response.json({ error: "Internal server error" }, { status: 500 });
+  }
+}