+++ app/api/v1/uploadthing/core.ts
@@ -1,0 +1,79 @@
+import { validateRequest } from "@/lib/auth/helper";
+import { prisma } from "@/lib/prisma";
+import streamServerClient from "@/lib/stream";
+import { createUploadthing, FileRouter } from "uploadthing/next";
+import { UploadThingError, UTApi } from "uploadthing/server";
+
+const f = createUploadthing();
+
+export const fileRouter = {
+  avatar: f({
+    image: { maxFileSize: "512KB" },
+  })
+    .middleware(async () => {
+      const { user } = await validateRequest();
+
+      if (!user) throw new UploadThingError("Unauthorized");
+
+      return { user };
+    })
+    .onUploadComplete(async ({ metadata, file }) => {
+      const oldImage = metadata.user.image;
+
+      if (oldImage) {
+        const key = oldImage.split(
+          `/a/${process.env.NEXT_PUBLIC_UPLOADTHING_APP_ID}/`,
+        )[1];
+
+        await new UTApi().deleteFiles(key);
+      }
+
+      const newImage = file.url.replace(
+        "/f/",
+        `/a/${process.env.NEXT_PUBLIC_UPLOADTHING_APP_ID}/`,
+      );
+
+      await Promise.all([
+        prisma.user.update({
+          where: { id: metadata.user.id },
+          data: {
+            image: newImage,
+          },
+        }),
+        streamServerClient.partialUpdateUser({
+          id: metadata.user.id,
+          set: {
+            image: newImage,
+          },
+        }),
+      ]);
+
+      return { avatarUrl: newImage };
+    }),
+  attachment: f({
+    image: { maxFileSize: "4MB", maxFileCount: 5 },
+    video: { maxFileSize: "64MB", maxFileCount: 5 },
+  })
+    .middleware(async () => {
+      const { user } = await validateRequest();
+
+      if (!user) throw new UploadThingError("Unauthorized");
+
+      return {};
+    })
+    .onUploadComplete(async ({ file }) => {
+      const media = await prisma.media.create({
+        data: {
+          url: file.url.replace(
+            "/f/",
+            `/a/${process.env.NEXT_PUBLIC_UPLOADTHING_APP_ID}/`,
+          ),
+          type: file.type.startsWith("image") ? "IMAGE" : "VIDEO",
+        },
+      });
+
+      return { mediaId: media.id };
+    }),
+} satisfies FileRouter;
+
+export type AppFileRouter = typeof fileRouter;