+++ next.config.js
@@ -9,6 +9,14 @@ const nextConfig = {
       'avatars.githubusercontent.com', // GitHub avatars
     ],
   },
+  async rewrites() {
+    return [
+      {
+        source: '/api/v1/:path*',
+        destination: '/api/v1/:path*',
+      },
+    ];
+  },
 };
 
 module.exports = nextConfig;